fix: some issues and code organization
